:art: 개별 뉴스레터 페이지 api 파일 구조 수정
