+++ src/views/newsletter/api/index.ts
@@ -1,0 +1,2 @@
+export * from "./newsletter";
+export * from "./subscriber";
