+++ src/views/newsletter/api/subscriber.ts
@@ -1,14 +1,3 @@
-export const fetchNewsletter = async () => {
-  if (process.env.NODE_ENV === "development") {
-    return {
-      name: "My Newsletter",
-      id: "1",
-      description: "This is my newsletter",
-      send_frequency: "daily",
-    } as Newsletter.Info;
-  }
-};
-
 export const fetchSubscribers = async () => {
   if (process.env.NODE_ENV === "development") {
     return [
